Restructure into src directory (#190)
* move into src directory

* use version file
