+++ .gitignore
@@ -85,5 +85,3 @@ notebooks/*
 
 # Ignore link to weather_plots in web/static/
 weather_plots
-
-src/
