+++ scripts/pyro_test_server.py
@@ -2,7 +2,7 @@
 """
 Script to run a Pyro test server.
 """
-from huntsman.utils.pyro.test_server import run_test_server
+from huntsman.pocs.utils.pyro.test_server import run_test_server
 
 
 if __name__ == "__main__":
