+++ setup.cfg
@@ -19,6 +19,13 @@ all_files = 1
 upload-dir = docs/_build/html
 show-response = 1
 
+[options]
+zip_safe = False
+packages = find:
+include_package_data = True
+package_dir =
+    =src
+
 [metadata]
 author = Huntsman Telephoto Array Team
 author_email = [Lee Spitler]
