+++ src/huntsman/pocs/camera/__init__.py
@@ -6,8 +6,8 @@ from pocs.camera import create_cameras_from_config as create_local_cameras
 from pocs.utils import error
 from pocs.utils import logger as logger_module
 
-from huntsman.camera.pyro import Camera as PyroCamera
-from huntsman.utils import load_config
+from huntsman.pocs.camera.pyro import Camera as PyroCamera
+from huntsman.pocs.utils import load_config
 
 
 def list_distributed_cameras(ns_host=None, logger=None):
