+++ src/huntsman/pocs/scheduler/dispatch.py
@@ -5,8 +5,8 @@ from pocs.scheduler.field import Field
 from pocs.utils import current_time
 from pocs.scheduler.observation import Observation
 
-from huntsman.scheduler.observation import DitheredObservation
-from huntsman.utils import dither
+from huntsman.pocs.scheduler.observation import DitheredObservation
+from huntsman.pocs.utils import dither
 
 
 class Scheduler(dispatch.Scheduler):
