+++ src/huntsman/pocs/tests/test_camera.py
@@ -22,7 +22,7 @@ from pocs.scheduler.observation import Observation
 from pocs.utils.images import fits as fits_utils
 from pocs.utils import error
 
-from huntsman.camera.pyro import Camera as PyroCamera
+from huntsman.pocs.camera.pyro import Camera as PyroCamera
 
 sys.excepthook = Pyro4.util.excepthook
