+++ src/huntsman/pocs/tests/test_observatory.py
@@ -7,8 +7,8 @@ from pocs.scheduler import create_scheduler_from_config
 from pocs.dome import create_dome_from_config
 from pocs.mount import create_mount_from_config
 
-from huntsman.camera import create_cameras_from_config
-from huntsman.observatory import HuntsmanObservatory as Observatory
+from huntsman.pocs.camera import create_cameras_from_config
+from huntsman.pocs.observatory import HuntsmanObservatory as Observatory
 
 
 @pytest.fixture(scope='function')
