+++ src/huntsman/pocs/utils/dither.py
