+++ src/huntsman/pocs/utils/hdr.py
@@ -1,4 +1,4 @@
-from huntsman.utils import dither
+from huntsman.pocs.utils import dither
 
 from astropy import units as u
 from astropy.coordinates import SkyCoord
